Address PR review: relocate doc sources, JSON options, exposure docs
Responds to Mike Alhayek's review on PR #125.

- move the documentation search tool-instance stack out of the MCP project
  (CrestApps.Core.AI.Mcp) into CrestApps.Core.AI under
  Tooling/Instances/Documentation. These are ordinary tool instance sources
  usable without the MCP server, so they no longer live in or depend on the
  MCP package. The doc HTTP client name moves to DocumentationToolConstants
  and the plain AddHttpClient registration matches the sibling
  HttpApiRequest source (no MCP-only resilience dependency)
- replace per-property [JsonPropertyName] attributes in the Algolia and
  search-index sources with a JsonSerializerOptions (web defaults / camelCase)
  passed to the serialize and deserialize calls
- skip building the allow-list in the MCP list and call handlers when
  ExposeAllTools is true
- add a Breaking Changes note to the 1.1.0 changelog: MCP servers no longer
  expose tools by default (behavior change from 1.0.0)
- reword the changelog so the documentation sources are described as tool
  instance sources (not a replacement for a never-shipped search_documentation
  tool, and not part of the MCP server builder)
- fold the standalone documentation-search page into the MCP server feature
  page and remove it from the sidebar
- clarify in the MCP server docs that ExposeAllTools ignores the Tools list
- relocate the documentation search tests to Core/Tools to match the move
