+++ src/CrestApps.Core.Docs/docs/changelog/1.1.0.md
@@ -19,6 +19,16 @@ page will be updated as changes land after 1.0.0.
 
 - upgrades the framework's dependency baseline, including YesSql 6.0 (new `ISession.SaveAsync` signature), the Model Context Protocol 2.0 packages, the GitHub Copilot SDK 1.0.8 (new `PermissionsApi.SetAllowAllAsync` mode-based API), Anthropic 12.39.0, OllamaSharp 5.4.30, the .NET 10.0.10 runtime/extension packages, and the `Microsoft.Extensions.AI` 10.8.3 packages
 
+## Breaking Changes
+
+- **MCP servers no longer expose tools by default.** In 1.0.0 a server registered with
+  `WithCrestAppsHandlers()` listed and invoked every non-hidden registered tool. In 1.1.0 tool exposure is
+  opt-in through the new `McpServerOptions` site settings: nothing is listed or callable until you either
+  add the tool or tool instance name to `McpServerOptions.Tools` or set `McpServerOptions.ExposeAllTools`
+  to `true`. Hosts upgrading from 1.0.0 that relied on the previous "expose everything" behavior must set
+  `ExposeAllTools = true` (or populate `Tools`) to keep exposing their tools. Prompts and resources are
+  unaffected and are still always registered
+
 ## MCP server tool exposure
 
 - reworks tool exposure into an opt-in allow-list driven by the `McpServerOptions` site settings. Nothing
@@ -32,16 +42,17 @@ page will be updated as changes land after 1.0.0.
 
 ## Documentation search tool instances
 
-- replaces the earlier single `search_documentation` tool with documentation search
-  [tool instance sources](../core/tool-instances.md), so a host exposes one callable search function per
-  documentation site it configures. Register the sources on the tool instances builder with
-  `AddDocumentationSearchSources()` (or the individual `AddSitemapDocumentationSource()`,
-  `AddSearchIndexDocumentationSource()`, and `AddAlgoliaDocumentationSource()` methods). Each configured
-  instance binds one site and surfaces as a distinct function the AI model can call, and the instances are
-  managed and persisted through the existing tool instance store (YesSql or Entity Framework Core) and UI.
-  The MVC and Blazor sample-host tool instance editors add source-specific field groups so operators can
-  configure a documentation site (base URL, sitemap or index URL, Algolia application/index/search-only
-  key, and per-instance result limits) directly from the create and edit forms
+- adds documentation search [tool instance sources](../core/tool-instances.md), so a host exposes one
+  callable search function per documentation site it configures. These are ordinary tool instance sources
+  registered on the tool instances builder with `AddDocumentationSearchSources()` (or the individual
+  `AddSitemapDocumentationSource()`, `AddSearchIndexDocumentationSource()`, and
+  `AddAlgoliaDocumentationSource()` methods), so they can be used with or without the MCP server. Each
+  configured instance binds one site and surfaces as a distinct function the AI model can call, and the
+  instances are managed and persisted through the existing tool instance store (YesSql or Entity Framework
+  Core) and UI. The MVC and Blazor sample-host tool instance editors add source-specific field groups so
+  operators can configure a documentation site (base URL, sitemap or index URL, Algolia
+  application/index/search-only key, and per-instance result limits) directly from the create and edit
+  forms
 - ships three documentation search strategies, each as its own source and settings model: the sitemap
   source crawls a site through its `sitemap.xml` (for example a public Docusaurus site such as
   `core.crestapps.com`), the search-index source downloads a prebuilt JSON search index (for example a
