+++ src/CrestApps.Core.Docs/sidebars.js
@@ -71,7 +71,6 @@ const sidebars = {
                         'mcp/client',
                         'mcp/resource-types',
                         'mcp/server',
-                        'mcp/documentation-search',
                     ],
                 },
                 {
