+++ src/Primitives/CrestApps.Core.AI/Tooling/Instances/Documentation/AlgoliaDocSearchSite.cs
@@ -1,4 +1,4 @@
-namespace CrestApps.Core.AI.Mcp.Documentation;
+namespace CrestApps.Core.AI.Tooling.Instances.Documentation;
 
 /// <summary>
 /// Describes a documentation site that is searchable through the Algolia DocSearch query API (the
