+++ src/Primitives/CrestApps.Core.AI/Tooling/Instances/Documentation/AlgoliaDocumentationToolSettings.cs
@@ -1,4 +1,4 @@
-namespace CrestApps.Core.AI.Mcp.Documentation;
+namespace CrestApps.Core.AI.Tooling.Instances.Documentation;
 
 /// <summary>
 /// The user-provided settings for an Algolia DocSearch documentation search tool instance. The settings
