+++ src/Primitives/CrestApps.Core.AI/Tooling/Instances/Documentation/CachingDocumentationSource.cs
@@ -1,4 +1,4 @@
-namespace CrestApps.Core.AI.Mcp.Documentation;
+namespace CrestApps.Core.AI.Tooling.Instances.Documentation;
 
 /// <summary>
 /// A base class for documentation sources that materialize a full <see cref="DocumentationCorpus"/>
