+++ src/Primitives/CrestApps.Core.AI/Tooling/Instances/Documentation/DocumentationSearchRequest.cs
@@ -1,4 +1,4 @@
-namespace CrestApps.Core.AI.Mcp.Documentation;
+namespace CrestApps.Core.AI.Tooling.Instances.Documentation;
 
 /// <summary>
 /// Represents a single documentation search query issued against an <see cref="IDocumentationSource"/>.
