+++ src/Primitives/CrestApps.Core.AI/Tooling/Instances/Documentation/DocumentationSearchToolFunction.cs
@@ -7,7 +7,7 @@ using Microsoft.Extensions.AI;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
-namespace CrestApps.Core.AI.Mcp.Documentation;
+namespace CrestApps.Core.AI.Tooling.Instances.Documentation;
 
 /// <summary>
 /// An <see cref="AIFunction"/> produced by a documentation search tool instance source. Each function is
