+++ src/Primitives/CrestApps.Core.AI/Tooling/Instances/Documentation/IDocumentationSource.cs
@@ -1,4 +1,4 @@
-namespace CrestApps.Core.AI.Mcp.Documentation;
+namespace CrestApps.Core.AI.Tooling.Instances.Documentation;
 
 /// <summary>
 /// Represents a searchable documentation knowledge base. Implement this interface to expose a custom
