+++ src/Primitives/CrestApps.Core.AI/Tooling/Instances/Documentation/IDocumentationSourceMaterializer.cs
@@ -1,4 +1,4 @@
-namespace CrestApps.Core.AI.Mcp.Documentation;
+namespace CrestApps.Core.AI.Tooling.Instances.Documentation;
 
 /// <summary>
 /// Caches the runtime <see cref="IDocumentationSource"/> materialized for each configured documentation
