+++ src/Primitives/CrestApps.Core.AI/Tooling/Instances/Documentation/SearchIndexDocumentationToolSource.cs
@@ -5,14 +5,15 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Extensions.Options;
 
-namespace CrestApps.Core.AI.Mcp.Documentation;
+namespace CrestApps.Core.AI.Tooling.Instances.Documentation;
 
 /// <summary>
-/// The built-in <see cref="IAIToolInstanceSource"/> that lets users configure documentation search through
-/// the hosted Algolia DocSearch query API (the search used by many Docusaurus sites). Each configured
-/// <see cref="AIToolInstance"/> binds one Algolia index; the AI model only supplies the search query.
+/// The built-in <see cref="IAIToolInstanceSource"/> that lets users configure documentation search over a
+/// single site that publishes a prebuilt JSON search index (for example a MkDocs Material
+/// <c>search_index.json</c>). Each configured <see cref="AIToolInstance"/> binds one site; the AI model
+/// only supplies the search query.
 /// </summary>
-public sealed class AlgoliaDocumentationToolSource : IAIToolInstanceSource
+public sealed class SearchIndexDocumentationToolSource : IAIToolInstanceSource
 {
     /// <summary>
     /// Creates the <see cref="DocumentationSearchToolFunction"/> bound to the supplied instance's settings.
@@ -23,9 +24,9 @@ public sealed class AlgoliaDocumentationToolSource : IAIToolInstanceSource
     {
         ArgumentNullException.ThrowIfNull(instance);
 
-        var settings = instance.TryGet<AlgoliaDocumentationToolSettings>(out var stored)
+        var settings = instance.TryGet<SearchIndexDocumentationToolSettings>(out var stored)
             ? stored
-            : new AlgoliaDocumentationToolSettings();
+            : new SearchIndexDocumentationToolSettings();
 
         var functionName = instance.GetFunctionName();
         var description = string.IsNullOrWhiteSpace(instance.Description)
@@ -34,23 +35,23 @@ public sealed class AlgoliaDocumentationToolSource : IAIToolInstanceSource
 
         return new DocumentationSearchToolFunction(functionName, description, instance, services =>
         {
-            var site = new AlgoliaDocSearchSite
+            var site = new DocumentationSearchIndexSite
             {
                 Name = string.IsNullOrWhiteSpace(instance.Name)
                     ? functionName
                     : instance.Name,
-                ApplicationId = settings.ApplicationId,
-                ApiKey = settings.ApiKey,
-                IndexName = settings.IndexName,
+                BaseUrl = settings.BaseUrl,
+                IndexUrl = settings.IndexUrl,
                 MaxResults = settings.MaxResults,
             };
 
             var options = services.GetService<IOptions<DocumentationSearchOptions>>()?.Value ?? new DocumentationSearchOptions();
             var httpClientFactory = services.GetRequiredService<IHttpClientFactory>();
-            var logger = services.GetService<ILoggerFactory>()?.CreateLogger<AlgoliaDocumentationSource>()
-                ?? NullLogger<AlgoliaDocumentationSource>.Instance;
+            var timeProvider = services.GetService<TimeProvider>() ?? TimeProvider.System;
+            var logger = services.GetService<ILoggerFactory>()?.CreateLogger<SearchIndexDocumentationSource>()
+                ?? NullLogger<SearchIndexDocumentationSource>.Instance;
 
-            return new AlgoliaDocumentationSource(site, options, httpClientFactory, logger);
+            return new SearchIndexDocumentationSource(site, options, httpClientFactory, timeProvider, logger);
         });
     }
 }
