+++ src/Primitives/CrestApps.Core.AI/Tooling/Instances/Documentation/SitemapDocumentationToolSource.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Extensions.Options;
 
-namespace CrestApps.Core.AI.Mcp.Documentation;
+namespace CrestApps.Core.AI.Tooling.Instances.Documentation;
 
 /// <summary>
 /// The built-in <see cref="IAIToolInstanceSource"/> that lets users configure documentation search over a
