+++ src/Startup/CrestApps.Core.Blazor.Web/Components/Pages/Tooling/ToolInstances/Create.razor
@@ -5,7 +5,7 @@
 @using CrestApps.Core.AI
 @using CrestApps.Core.AI.Tooling
 @using CrestApps.Core.AI.Tooling.Instances
-@using CrestApps.Core.AI.Mcp.Documentation
+@using CrestApps.Core.AI.Tooling.Instances.Documentation
 @using CrestApps.Core.Blazor.Web.ViewModels
 @using CrestApps.Core.Services
 @using Microsoft.Extensions.Options
